+++ content_service/src/main/java/api/user/service/UserContentService.java
@@ -1,0 +1,4 @@
+package application.service.user;
+
+public class UserContentService {
+}
