--- content_service/src/main/java/application/port/in/UserContentService.java
@@ -1,4 +1,0 @@
-package application.port.in;
-
-public interface UserContentService {
-}
