--- content_service/src/main/java/application/service/user/UserContentServiceImpl.java
@@ -1,14 +1,0 @@
-package application.service.user;
-
-import application.port.in.UserContentService;
-import application.port.out.UserContentRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-
-@Service
-@RequiredArgsConstructor
-public class UserContentServiceImpl implements UserContentService {
-
-    private final UserContentRepository userContentRepository;
-}
