+++ content_service/src/main/java/domain/define/BaseEntity.java
