+++ content_service/src/main/java/domain/define/user/constant/ContentInfo.java
